+++ FeetOfAzeroth/CHANGELOG.md
@@ -14,8 +14,22 @@
 - ajout : switch entre yards et mètres
 - bugfix : petit soucis avec les arrondis
 
+### 0.3 - August 11,2026
+- GUI review
+- ajout : affichage de la vitesse du joueur
+- ajout : le bouton yd/m converti maintenant la distance ET la vitesse
 
+### 0.4 - August 12.2026
+- ajout : bouton stats
+- ajout : fenêtre stats
+- ajout : bouton reset pour réinitialiser
+- ajout : bouton achivevement pour voir les succès
+- ajout : 3 succès
+- ajout : fenêtre des succès
 
-TODO :
+## TODO 
+
 - sauvegarder l'unité choisie
+- bug si on se téléporte : la distance ne devrait pas être comptée
+- mot sur le bouton achievement qui est trop long